Update transaction handler test paths - fix failing test
