+++ backend/banking/utils.py
@@ -1,37 +1,21 @@
 """
-banking/utils.py — Transaction import utilities and domain-level business logic.
+backend/banking/utils.py — Account-detection utilities.
 
 This module contains non-model business logic used during transaction
-imports, including:
-
-- inferring an account type from uploaded CSV filenames
-- inserting (upserting) transactions while preserving user-managed fields
+imports: inferring an account type from uploaded CSV filenames.
 
 Account type detection relies on canonical handler keys defined in
-banking.constants.HandlerKeys. These keys are system-defined,
-seeded via data migrations, and resolved at runtime through the account
-handler registry.
+banking.constants.HandlerKeys. These keys are system-defined, seeded via
+data migrations, and resolved at runtime through the account handler
+registry.
 
-This module intentionally does not contain parsing logic or persistence
-rules beyond transaction insertion; CSV parsing and normalization are
-handled by account handlers, while referential integrity is enforced at
-the model layer.
-
-Note: Transaction itself remains a transactions-app model — only the
-account-detection and upsert logic lives here, since it depends on
-banking.models.Account and the handler registry.
+Transaction persistence (upsert_transactions) has moved to
+transactions/utils.py — banking owns account/handler concerns only;
+anything that creates or queries Transaction rows belongs in the
+transactions app.
 """
 
-import logging
-
-import pandas as pd
-
 from banking.constants import HandlerKeys
-from banking.models import Account
-from transactions.models import Transaction  # Transaction model still lives in transactions app
-
-logger = logging.getLogger(__name__)
-
 
 # ── Account detection ─────────────────────────────────────────────────────────
 
@@ -62,75 +46,3 @@ def detect_account_type(filename: str) -> str | None:
         if substring in filename:
             return handler_key
     return None
-
-
-# ── Transaction upsert ────────────────────────────────────────────────────────
-
-
-def upsert_transactions(df: pd.DataFrame, account: Account) -> dict:
-    """
-    Insert new transactions from a DataFrame, skipping duplicates.
-    Deduplication is scoped to the account via dedupe_hash — preventing
-    both accidental duplicates and cross-tenant ID collisions.
-    Labels, category, and additional_labels are never overwritten on re-import.
-
-    Args:
-        df:      Cleaned DataFrame from a handler's process() method.
-        account: The Account instance transactions belong to.
-
-    Returns:
-        dict with keys: inserted, skipped, total.
-    """
-    if df.empty:
-        return {'inserted': 0, 'skipped': 0, 'total': 0}
-
-    total = len(df)
-
-    # De-dupe within the incoming batch — banks occasionally export the same row twice
-    df = df.drop_duplicates(subset=['dedupe_hash'])
-
-    # Extract all dedupe hashes from the DataFrame
-    incoming_hashes = df['dedupe_hash'].tolist()
-
-    # Fetch existing transaction dedupe hashes in one query
-    existing_hashes = set(
-        Transaction.objects.filter(account=account, dedupe_hash__in=incoming_hashes).values_list(
-            'dedupe_hash', flat=True
-        )
-    )
-
-    # Build list of new transactions to insert
-    new_transactions = []
-    for row in df.itertuples(index=False):
-        if row.dedupe_hash not in existing_hashes:
-            new_transactions.append(
-                Transaction(
-                    dedupe_hash=row.dedupe_hash,
-                    raw_data=row.raw_data,
-                    date=row.Date.date() if hasattr(row.Date, 'date') else row.Date,
-                    concept=row.Concept,
-                    amount=row.Amount,
-                    label=None,
-                    category=None,
-                    additional_labels=None,
-                    source=Transaction.Source.IMPORT,
-                    account=account,
-                )
-            )
-
-    # Bulk insert new transactions
-    if new_transactions:
-        Transaction.objects.bulk_create(new_transactions, ignore_conflicts=True)
-
-    inserted = Transaction.objects.filter(
-        account=account,
-        dedupe_hash__in=[t.dedupe_hash for t in new_transactions],
-    ).count()
-    skipped = total - inserted
-
-    logger.info(
-        f"Upsert complete for account '{account.name}' — "
-        f'inserted: {inserted}, skipped: {skipped}, total: {total}'
-    )
-
-    return {'inserted': inserted, 'skipped': skipped, 'total': total}
