--- backend/tests/transactions/handlers/__init__.py
@@ -1,1 +1,0 @@
-# tests/handlers/__init__.py
