--- backend/transactions/handlers/__init__.py
@@ -1,1 +1,0 @@
-# handlers/__init__.py
